Add missing *WithIndex instances
