+++ optics-core/src/Optics/Internal/Indexed.hs
@@ -4,13 +4,9 @@
 module Optics.Internal.Indexed where
 
 import Control.Applicative
-import Control.Applicative.Backwards
-import Control.Monad.Trans.Identity
-import Control.Monad.Trans.Reader
 import Data.Functor.Compose
 import Data.Functor.Identity
 import Data.Functor.Product
-import Data.Functor.Reverse
 import Data.Functor.Sum
 import Data.Ix
 import Data.List.NonEmpty
@@ -189,35 +185,6 @@ ifor = flip itraverse
 ----------------------------------------
 -- Instances
 
--- Backwards
-
-instance FunctorWithIndex i f => FunctorWithIndex i (Backwards f) where
-  imap f  = Backwards . imap f . forwards
-  {-# INLINE imap #-}
-
-instance FoldableWithIndex i f => FoldableWithIndex i (Backwards f) where
-  ifoldMap f = ifoldMap f . forwards
-  {-# INLINE ifoldMap #-}
-
-instance TraversableWithIndex i f => TraversableWithIndex i (Backwards f) where
-  itraverse f = fmap Backwards . itraverse f . forwards
-  {-# INLINE itraverse #-}
-
--- Reverse
-
-instance FunctorWithIndex i f => FunctorWithIndex i (Reverse f) where
-  imap f = Reverse . imap f . getReverse
-  {-# INLINE imap #-}
-
-instance FoldableWithIndex i f => FoldableWithIndex i (Reverse f) where
-  ifoldMap f = getDual . ifoldMap (\i -> Dual #. f i) . getReverse
-  {-# INLINE ifoldMap #-}
-
-instance TraversableWithIndex i f => TraversableWithIndex i (Reverse f) where
-  itraverse f =
-    fmap Reverse . forwards . itraverse (\i -> Backwards . f i) . getReverse
-  {-# INLINE itraverse #-}
-
 -- Identity
 
 instance FunctorWithIndex () Identity where
@@ -322,7 +289,11 @@ instance FunctorWithIndex Int IntMap.IntMap where
   {-# INLINE imap #-}
 instance FoldableWithIndex Int IntMap.IntMap where
   ifoldMap = IntMap.foldMapWithKey
+  ifoldr   = IntMap.foldrWithKey
+  ifoldl'  = IntMap.foldlWithKey' . flip
   {-# INLINE ifoldMap #-}
+  {-# INLINE ifoldr #-}
+  {-# INLINE ifoldl' #-}
 instance TraversableWithIndex Int IntMap.IntMap where
   itraverse = IntMap.traverseWithKey
   {-# INLINE itraverse #-}
@@ -334,7 +305,11 @@ instance FunctorWithIndex k (Map.Map k) where
   {-# INLINE imap #-}
 instance FoldableWithIndex k (Map.Map k) where
   ifoldMap = Map.foldMapWithKey
+  ifoldr   = Map.foldrWithKey
+  ifoldl'  = Map.foldlWithKey' . flip
   {-# INLINE ifoldMap #-}
+  {-# INLINE ifoldr #-}
+  {-# INLINE ifoldl' #-}
 instance TraversableWithIndex k (Map.Map k) where
   itraverse = Map.traverseWithKey
   {-# INLINE itraverse #-}
@@ -393,20 +368,6 @@ instance (TraversableWithIndex i f, TraversableWithIndex j g
   itraverse q (InR ga) = InR <$> itraverse (q . Right) ga
   {-# INLINE itraverse #-}
 
--- IdentityT
-
-instance FunctorWithIndex i m => FunctorWithIndex i (IdentityT m) where
-  imap f (IdentityT m) = IdentityT $ imap f m
-  {-# INLINE imap #-}
-
-instance FoldableWithIndex i m => FoldableWithIndex i (IdentityT m) where
-  ifoldMap f (IdentityT m) = ifoldMap f m
-  {-# INLINE ifoldMap #-}
-
-instance TraversableWithIndex i m => TraversableWithIndex i (IdentityT m) where
-  itraverse f (IdentityT m) = IdentityT <$> itraverse f m
-  {-# INLINE itraverse #-}
-
 -- Product
 
 instance (FunctorWithIndex i f, FunctorWithIndex j g
@@ -425,12 +386,6 @@ instance (TraversableWithIndex i f, TraversableWithIndex j g
   itraverse f (Pair a b) =
     Pair <$> itraverse (f . Left) a <*> itraverse (f . Right) b
   {-# INLINE itraverse #-}
-
--- ReaderT
-
-instance FunctorWithIndex i m => FunctorWithIndex (e, i) (ReaderT e m) where
-  imap f (ReaderT m) = ReaderT $ \k -> imap (f . (,) k) (m k)
-  {-# INLINE imap #-}
 
 -- Tree
